# app-annonces/src/swagger.yml
openapi: 3.0.0
info:
  title: API Plateforme d'Annonces
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement
  - url: https://api.annonces.com/api
    description: Serveur de production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu après connexion
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nom:
          type: string
          example: "Jean Dupont"
        email:
          type: string
          format: email
          example: "jean@example.com"
        role:
          type: string
          enum: [annonceur, acheteur, administrateur]
          example: "annonceur"
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Erreur de validation"
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentification
      summary: Inscription d'un nouvel utilisateur
      description: Créer un nouveau compte utilisateur (annonceur ou acheteur)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nom
                - email
                - password
              properties:
                nom:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "Jean Dupont"
                email:
                  type: string
                  format: email
                  example: "jean@example.com"
                password:
                  type: string
                  minLength: 6
                  example: "motdepasse123"
                role:
                  type: string
                  enum: [annonceur, acheteur]
                  default: annonceur
                  example: "annonceur"
      responses:
        '201':
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilisateur créé avec succès"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Erreur de validation ou email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentification
      summary: Connexion d'un utilisateur
      description: Authentifier un utilisateur et obtenir un token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "jean@example.com"
                password:
                  type: string
                  example: "motdepasse123"
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Connexion réussie"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email ou mot de passe incorrect"
  /auth/me:
    get:
      tags:
        - Authentification
      summary: Récupérer son profil
      description: Obtenir les informations de l'utilisateur connecté
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token manquant"
  /auth/update-profile:
    put:
      tags:
        - Authentification
      summary: Mettre à jour son profil
      description: Modifier les informations de son compte
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nom:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: "Jean Martin"
                email:
                  type: string
                  format: email
                  example: "jean.martin@example.com"
                password:
                  type: string
                  minLength: 6
                  example: "nouveaumotdepasse"
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Profil mis à jour avec succès"
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
...